CI: there is only 1 seed file and command. We parse CLI args to determine what to run based on environment
