+++ .github/workflows/deploy-production.yml
@@ -25,7 +25,7 @@ jobs:
 
       - name: Run Prisma SeedMain
         run: |
-          npx prisma db seedMain
+          npx prisma db seed -- --environment production
 
       - name: Install Vercel CLI
         run: npm install --global vercel@latest
