+++ package.json
@@ -19,8 +19,7 @@
     "pre-commit": "lint-staged && npm test"
   },
   "prisma": {
-    "seed": "tsx prisma/seed.ts",
-    "seedMain": "tsx prisma/seed_main.ts"
+    "seed": "tsx prisma/seed.ts"
   },
   "engines": {
     "node": "18.x"
